frontend/requirements.txt
tornado
websockets
python-dotenv
grpcio 
grpcio-tools